chore: bump nextjs version and enable turbopack
