+++ package.json
@@ -6,10 +6,12 @@
         "node": ">=21.1.0 <22.0.0 || ^18.18.0 || ^20.9.0"
     },
     "scripts": {
-        "dev": "next dev",
-        "dev:tunnel": "ngrok http 3000",
+        "dev": "next dev --turbo",
+        "dev:fallback": "next dev",
         "build": "next build",
+        "build:fallback": "next build --turbo",
         "start": "next start",
+        "start:fallback": "next start --turbo",
         "lint": "next lint",
         "knip": "knip",
         "format": "prettier --write .",
@@ -58,10 +60,10 @@
         "jsqr": "^1.4.0",
         "lottie-react": "^2.4.0",
         "multicoin-address-validator": "^0.5.22",
-        "next": "^14.2.18",
-        "react": "^18.3.1",
+        "next": "14.1.0",
+        "react": "^18.2.0",
         "react-csv": "^2.2.2",
-        "react-dom": "^18.3.1",
+        "react-dom": "^18.2.0",
         "react-fast-marquee": "^1.6.5",
         "react-ga4": "^2.1.0",
         "react-hook-form": "^7.53.2",
