+++ src/app/api/peanut/user/get-user-from-cookie/route.ts
@@ -1,10 +1,10 @@
-import { cookies } from 'next/headers'
-import { NextRequest, NextResponse } from 'next/server'
 import * as consts from '@/constants'
 import { fetchWithSentry } from '@/utils'
+import { cookies } from 'next/headers'
+import { NextRequest, NextResponse } from 'next/server'
 
 export async function GET(_request: NextRequest) {
-    const cookieStore = cookies()
+    const cookieStore = await cookies()
     const token = cookieStore.get('jwt-token')
     const apiKey = process.env.PEANUT_API_KEY
