+++ src/app/api/peanut/user/logout-user/route.ts
@@ -1,8 +1,8 @@
-import { NextRequest, NextResponse } from 'next/server'
 import { cookies } from 'next/headers'
+import { NextRequest, NextResponse } from 'next/server'
 
 export async function GET(request: NextRequest) {
-    const cookieStore = cookies()
+    const cookieStore = await cookies()
     const token = cookieStore.get('jwt-token')
 
     if (!token) {
